planner/requirements.txt
paho.mqtt
python-dotenv